requirements.txt
azure-functions
request